요구사항에 따른 User 클래스 관련 변경
- User인스턴스 저장 자료구조 변경
HashMap -> ArrayList

- User id 멤버변수 타입 변경
Long -> int

+++ src/main/java/com/kakao/cafe/repository/UserRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 public interface UserRepository {
 
     User save(User user);
-    Optional<User> findById(Long id);
+    Optional<User> findById(int id);
     Optional<User> findByUserId(String userId);
     List<User> findAll();
